Change molecules to Material 3
