+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/molecule/ContentLoadingErrorView.kt
@@ -1,17 +1,10 @@
 package app.k9mail.core.ui.compose.designsystem.molecule
 
 import androidx.compose.animation.AnimatedContent
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.tooling.preview.Preview
-import app.k9mail.core.ui.compose.designsystem.atom.text.TextSubtitle1
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
 
 @Composable
 fun ContentLoadingErrorView(
@@ -43,44 +36,4 @@ enum class ContentLoadingErrorState {
     Loading,
     Content,
     Error,
-}
-
-@Composable
-@Preview(showBackground = true)
-internal fun ContentLoadingErrorViewPreview() {
-    PreviewWithThemes {
-        val state = remember {
-            mutableStateOf(ContentLoadingErrorState.Loading)
-        }
-
-        ContentLoadingErrorView(
-            state = state.value,
-            modifier = Modifier
-                .clickable {
-                    when (state.value) {
-                        ContentLoadingErrorState.Loading -> {
-                            state.value = ContentLoadingErrorState.Content
-                        }
-
-                        ContentLoadingErrorState.Content -> {
-                            state.value = ContentLoadingErrorState.Error
-                        }
-
-                        ContentLoadingErrorState.Error -> {
-                            state.value = ContentLoadingErrorState.Loading
-                        }
-                    }
-                }
-                .fillMaxSize(),
-            error = {
-                TextSubtitle1(text = "Error")
-            },
-            loading = {
-                TextSubtitle1(text = "Loading...")
-            },
-            content = {
-                TextSubtitle1(text = "Content")
-            },
-        )
-    }
 }
