+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/molecule/ErrorView.kt
@@ -8,20 +8,18 @@ import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.R
 import app.k9mail.core.ui.compose.designsystem.atom.button.ButtonText
+import app.k9mail.core.ui.compose.designsystem.atom.icon.Icon
 import app.k9mail.core.ui.compose.designsystem.atom.icon.Icons
-import app.k9mail.core.ui.compose.designsystem.atom.text.TextBody2
-import app.k9mail.core.ui.compose.designsystem.atom.text.TextSubtitle1
-import app.k9mail.core.ui.compose.theme.MainTheme
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
+import app.k9mail.core.ui.compose.designsystem.atom.text.TextBodyMedium
+import app.k9mail.core.ui.compose.designsystem.atom.text.TextHeadlineSmall
+import app.k9mail.core.ui.compose.theme2.MainTheme
 
 @Composable
 fun ErrorView(
@@ -40,19 +38,18 @@ fun ErrorView(
         Column(
             modifier = Modifier
                 .fillMaxWidth()
-                .padding(MainTheme.spacings.double),
+                .padding(MainTheme.spacings.triple),
         ) {
             Column(
                 modifier = Modifier.fillMaxWidth(),
                 horizontalAlignment = Alignment.CenterHorizontally,
-                verticalArrangement = Arrangement.spacedBy(MainTheme.spacings.default),
             ) {
                 Icon(
                     imageVector = Icons.Filled.error,
-                    contentDescription = null,
                     tint = MainTheme.colors.error,
                 )
-                TextSubtitle1(
+                Spacer(modifier = Modifier.height(MainTheme.spacings.double))
+                TextHeadlineSmall(
                     text = title,
                     textAlign = TextAlign.Center,
                 )
@@ -60,13 +57,13 @@ fun ErrorView(
 
             if (message != null) {
                 Spacer(modifier = Modifier.height(MainTheme.spacings.double))
-                TextBody2(
+                TextBodyMedium(
                     text = message,
                     modifier = Modifier.fillMaxWidth(),
                 )
             }
             if (onRetry != null) {
-                Spacer(modifier = Modifier.height(MainTheme.spacings.default))
+                Spacer(modifier = Modifier.height(MainTheme.spacings.triple))
                 Row(
                     modifier = Modifier.fillMaxWidth(),
                     horizontalArrangement = Arrangement.End,
@@ -78,49 +75,5 @@ fun ErrorView(
                 }
             }
         }
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun ErrorViewPreview() {
-    PreviewWithThemes {
-        ErrorView(
-            title = "Error",
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun ErrorViewWithMessagePreview() {
-    PreviewWithThemes {
-        ErrorView(
-            title = "Error",
-            message = "Something went wrong.",
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun ErrorViewWithRetryPreview() {
-    PreviewWithThemes {
-        ErrorView(
-            title = "Error",
-            onRetry = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun ErrorViewWithRetryAndMessagePreview() {
-    PreviewWithThemes {
-        ErrorView(
-            title = "Error",
-            message = "Something went wrong.",
-            onRetry = {},
-        )
     }
 }
