+++ core/ui/compose/designsystem/src/main/kotlin/app/k9mail/core/ui/compose/designsystem/molecule/input/EmailAddressInput.kt
@@ -5,10 +5,8 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import app.k9mail.core.ui.compose.designsystem.R
 import app.k9mail.core.ui.compose.designsystem.atom.textfield.TextFieldOutlinedEmailAddress
-import app.k9mail.core.ui.compose.theme.PreviewWithThemes
 
 @Composable
 fun EmailAddressInput(
@@ -31,27 +29,6 @@ fun EmailAddressInput(
             isEnabled = isEnabled,
             hasError = errorMessage != null,
             modifier = Modifier.fillMaxWidth(),
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun EmailAddressInputPreview() {
-    PreviewWithThemes {
-        EmailAddressInput(
-            onEmailAddressChange = {},
-        )
-    }
-}
-
-@Preview(showBackground = true)
-@Composable
-internal fun EmailAddressInputWithErrorPreview() {
-    PreviewWithThemes {
-        EmailAddressInput(
-            onEmailAddressChange = {},
-            errorMessage = "Email address error",
         )
     }
 }
